# specs/001-photo-story-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 二十年前与现在体验接口
  version: 1.0.0
  description: |
    接收已获明确授权的单张现场照片，在内存中完成示例匹配、
    可见非敏感差异分析和虚构故事生成。现场照片默认不留存。
    内部人物库可由多人合影构建，但 real 模式要求合影全员授权；
    匹配仍只返回一个具体人物，因此公开响应结构不随 v2 人物库变化。
servers:
  - url: /api
    description: 与 H5 同源的接口
paths:
  /health:
    get:
      summary: 检查服务健康状态
      responses:
        "200":
          description: 服务可用，不包含密钥或人物数据
          content:
            application/json:
              schema:
                type: object
                required: [status, providerMode]
                properties:
                  status:
                    type: string
                    const: ok
                  providerMode:
                    type: string
                    enum: [mock, real]
  /experience:
    post:
      summary: 创建一次不持久化的照片故事体验
      description: |
        图片只在本请求内存中处理，请求结束后尽力清零缓冲区。
        默认业务匹配阈值为 0.6；低于阈值不返回候选人物结论。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo, consent]
              properties:
                photo:
                  type: string
                  format: binary
                  description: JPEG、PNG 或 WebP，最大 6 MiB
                consent:
                  type: string
                  const: "true"
                  description: 用户已在本次体验中主动授权
                demoCase:
                  type: string
                  enum: [success, no-face, multiple-faces, unmatched, provider-error]
                  description: 仅 mock 模式用于演示状态
      responses:
        "200":
          description: 达到阈值并成功生成体验结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceResult"
        "400":
          description: 缺少授权、文件或图片格式无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: 图片超过 6 MiB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: 无人脸、多人脸或低于匹配阈值
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 未预期服务错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: provider 暂时失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /people/{personId}/photo:
    get:
      summary: 读取成功匹配人物的私有照片
      parameters:
        - in: path
          name: personId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 私有且不可缓存的人物照片
          headers:
            Cache-Control:
              schema:
                type: string
                const: private, no-store
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
            image/svg+xml: {}
        "403":
          description: 缺少、无效、过期或人物不匹配的照片 grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 人物或照片不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ExperienceResult:
      type: object
      required: [requestId, retention, match, differences, story]
      properties:
        requestId:
          type: string
        retention:
          type: string
          description: 现场照片保留说明
        match:
          type: object
          required: [matched, score, threshold, confidence, person]
          properties:
            matched:
              type: boolean
              const: true
            score:
              type: number
              minimum: 0
              maximum: 1
            threshold:
              type: number
              minimum: 0.5
              maximum: 0.99
            confidence:
              type: string
              enum: [high, medium]
            person:
              $ref: "#/components/schemas/Person"
        differences:
          type: array
          items:
            $ref: "#/components/schemas/VisibleDifference"
        story:
          $ref: "#/components/schemas/FictionStory"
    Person:
      type: object
      required: [id, displayName, oldPhotoUrl, sourceNote, faceBox]
      properties:
        id:
          type: string
        displayName:
          type: string
        oldPhotoUrl:
          type: string
        sourceNote:
          type: string
        faceBox:
          description: 匹配人物在旧照中的归一化位置；旧版占位人物可能没有坐标。
          oneOf:
            - type: object
              required: [left, top, width, height]
              properties:
                left:
                  type: number
                  minimum: 0
                  maximum: 1
                top:
                  type: number
                  minimum: 0
                  maximum: 1
                width:
                  type: number
                  exclusiveMinimum: 0
                  maximum: 1
                height:
                  type: number
                  exclusiveMinimum: 0
                  maximum: 1
            - type: "null"
    VisibleDifference:
      type: object
      required: [category, description]
      properties:
        category:
          type: string
          enum: [hairstyle, clothing, expression, accessory]
        description:
          type: string
    FictionStory:
      type: object
      required: [label, title, content, disclaimer]
      properties:
        label:
          type: string
          const: AI 创作/虚构
        title:
          type: string
        content:
          type: string
        disclaimer:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, explanation, retryable, requestId]
          properties:
            code:
              type: string
              enum:
                - CONSENT_REQUIRED
                - PHOTO_REQUIRED
                - INVALID_IMAGE
                - INVALID_REQUEST
                - PHOTO_TOO_LARGE
                - NO_FACE
                - MULTIPLE_FACES
                - MATCH_BELOW_THRESHOLD
                - PHOTO_GRANT_REQUIRED
                - PERSON_PHOTO_NOT_FOUND
                - RATE_LIMITED
                - PROVIDER_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
            explanation:
              type: string
            retryable:
              type: boolean
            requestId:
              type: string
            details:
              type: object
              additionalProperties: true